Embeddings/Architectures/Embeddings with Azure Functions/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
faiss-cpu
numpy
pandas
azure-data-tables
openai